% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{code}
\alias{code}
\title{Code}
\usage{
code(x, file, ...)
}
\arguments{
\item{x}{Block object.}

\item{file}{File (\link{new_file}) to use for context.}

\item{...}{Ignored.}
}
\description{
Generate code for a blockr block
}
