% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{content}
\alias{content}
\title{Content}
\usage{
content(x, ...)
}
\arguments{
\item{x}{Output of \code{\link[=new_file]{new_file()}}}

\item{...}{Ignored.}
}
\description{
Generate content for a \code{\link[=new_file]{new_file()}}.
}
