% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gzip.R
\name{export_gzip}
\alias{export_gzip}
\title{Export to set of R Scripts}
\usage{
export_gzip(
  file,
  workspace = blockr::get_workspace(),
  to_copy = c(),
  style = FALSE
)
}
\arguments{
\item{file}{The file to export to}

\item{workspace}{The workspace to export}

\item{to_copy}{A vector of package names to copy the functions from (if found).}

\item{style}{Whether to run \code{\link[styler:style_file]{styler::style_file()}} on the generated code.}
}
\description{
Export a workspace to a collection R scripts.
}
