% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front-matter.R
\name{front_matter}
\alias{front_matter}
\title{Front matter}
\usage{
front_matter(x, ...)
}
\arguments{
\item{x}{Output of \code{\link[=new_file]{new_file()}}}

\item{...}{Ignored.}
}
\description{
Generate the front matter for a \code{\link[=new_file]{new_file()}}.
}
