% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{new_file}
\alias{new_file}
\title{New file}
\usage{
new_file(
  file,
  ...,
  title = "Blockr workspace",
  workspace = blockr::get_workspace(),
  class
)
}
\arguments{
\item{file}{Path to file to generate.}

\item{...}{Any additional attribute.}

\item{title}{Title of the file.}

\item{workspace}{Blockr workspace as returned by \code{\link[blockr:set_workspace]{blockr::get_workspace()}}.}

\item{class}{Class of the new file.}
}
\description{
Create a new file
}
