% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-write.R
\name{post_write}
\alias{post_write}
\title{Post Write}
\usage{
post_write(x, ...)
}
\arguments{
\item{x}{Output of \code{\link[=new_file]{new_file()}}}

\item{...}{Ignored.}
}
\description{
Run callback after the (\link{content}) for a \code{\link[=new_file]{new_file()}} as been written.
}
