% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\title{Render}
\usage{
render(x, ...)
}
\arguments{
\item{x}{Output of \code{\link[=new_file]{new_file()}}}

\item{...}{Ignored.}
}
\description{
Render a \code{\link[=new_file]{new_file()}}, used to render Rmarkdown files.
}
