% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write}
\alias{write}
\title{Write}
\usage{
write(x, ...)
}
\arguments{
\item{x}{Output of \code{\link[=new_file]{new_file()}}}

\item{...}{Ignored.}
}
\description{
Write generated content (\link{content}) for a \code{\link[=new_file]{new_file()}}.
}
