% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{code_fence}
\alias{code_fence}
\title{Code Fence}
\usage{
code_fence(x, name, code, ...)
}
\arguments{
\item{x}{Stack object.}

\item{name}{Name of the stack.}

\item{code}{Code generated for the stack.}

\item{...}{Ignored.}
}
\description{
Generate code fence for a blockr stack
}
