% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R, R/markdown.R
\name{export_markdown}
\alias{export_markdown}
\title{Export to HTML}
\usage{
export_markdown(
  file,
  chunks = c("stack", "block"),
  workspace = blockr::get_workspace()
)

export_markdown(
  file,
  chunks = c("stack", "block"),
  workspace = blockr::get_workspace()
)
}
\arguments{
\item{file}{The file to export to}

\item{chunks}{Whether code chunks represent entire stacks or individual blocks.}

\item{workspace}{The workspace to export}
}
\description{
Export a workspace to HTML via Rmarkdown.

Export a workspace to markdown.
}
