% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{export_markdown}
\alias{export_markdown}
\title{Export to Markdown}
\usage{
export_markdown(
  file,
  chunks = c("stack", "block"),
  workspace = blockr::get_workspace(),
  to_copy = c()
)
}
\arguments{
\item{file}{The file to export to}

\item{chunks}{Whether code chunks represent entire stacks or individual blocks.}

\item{workspace}{The workspace to export}

\item{to_copy}{A vector of package names to copy the functions from (if found).}
}
\description{
Export a workspace to markdown.
}
