% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{export_rmarkdown}
\alias{export_rmarkdown}
\title{Export to RMarkdown}
\usage{
export_rmarkdown(
  file,
  workspace = blockr::get_workspace(),
  to_copy = c(),
  style = FALSE,
  ...,
  output = "html_document"
)
}
\arguments{
\item{file}{The file to export to}

\item{workspace}{The workspace to export}

\item{to_copy}{A vector of package names to copy the functions from (if found).}

\item{style}{Whether to run \code{\link[styler:style_file]{styler::style_file()}} on the generated code.}

\item{...}{Passed to \code{\link[=new_file]{new_file()}}}

\item{output}{Output file to use in the front matter.}
}
\description{
Export a workspace to Rmarkdown.
}
