% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{export_rmarkdown}
\alias{export_rmarkdown}
\title{Export a RMarkdown file}
\usage{
export_rmarkdown(
  file,
  chunks = c("stack", "block"),
  workspace = blockr::get_workspace()
)
}
\arguments{
\item{file}{The file to export to}

\item{chunks}{Whether code chunks represent entire stacks or individual blocks.}

\item{workspace}{The workspace to export}
}
\description{
Export a RMarkdown file
}
